requirements.txt
tensorflow_gpu==2.8.0
numpy==1.18.5
tqdm==4.54.1
opencv-python==4.7.0.68
